Skills system added
It's now possible to add skills as resources and use them in combat.
Some changes:

- Initial stats now have mana
- A combat action for skills has been added
- A really simple menu to select which action to perform in the given
turn
- Resource for skills
- Skills can be executed on a given chance

What can be improved:
- As of now, the skills have a `base_damage` which is added on top of the
already calculated damage in the `Hit` script. We might want to use the
strength as a multiplier to increase the overall damage of the skill,
for instance, instead of just adding them together.

- The menu for selecting the action is really simple and the UX is bad
to say the least. I'll create a separate issue for this.

- Skills are only available as a form of attack, i.e. you can't create
healing skill with the current system. But it can be extended to do so.

- They are not specific to jobs. For instance, both Dan and Blue are
fighters, but you'll have to add skills to them separately. Though, this
is most due to how they were created. We should probably create a
separate FighterJob and add the skills to them, this would fix the
problem imo.

Closes #16 and #15