Concept arts integrated into the combat system
The only "missing" point is converting the game from 720p to 1080p. For
what I can tell, it's already 1080p, but it's being tested as 720p.

I had to downscale the sprites for the battlers, they were occuping the
whole screen.

For the platforms, I decided to only hide the sprite:
- The platforms might be used in a different area theme in the future
- The characters already have their own shadow
- Shadows have different sizes depending on what battler is "casting" it

Closes #53