Added monster drops
Monsters can now be configured to drop items based on a resource (MonsterDrops.gd).
The usability is not the greatest, but that's due to not being able to export custom types. 
There were two possible approaches that I saw:
1. An `Array` holding dictionaries for each item that can be dropped along with the data needed for the drop behavior.
2.  Create a custom resource that would encapsulate all the data for the items that can be dropped from a monster: The item itself, min and max quantities, and the chance. 

I decided to go with the first one.

ps: The items are _NOT_ being added to the player's inventory . This is just the functionality needed for the monsters to start dropping items.

Closes #59