Refactor combat to support multiple targets and add the circular menu
Main changes:

- Simplified combat actions. Action handle the logic that relates to
their name. E.g. the attack-specific code is now in Attack.gd and not in
its parent class. Removed unused parameters. Removed actions selecting
their own target: this breaks the single responsibility principle (and
made the action code break when I changed stuff in the interface)
- Removed the old school UI and action_list code in favor of the
circular menu
- Almost all functions that dealed with one target now work with an
array of targets instead, even if the player only selects one target
- All commands should support working with multiple targets as a result

A lot of the changes I needed to make to accomodate for either feature
broke the game, so that's why I'm packing all the changes in a single
commit.