Characters dialogue database and portraits added
I opted to serve the textures with a function in the database that handles the fetching of the desired texture. It looked more natural, easier and more natural than to add this logic to the Dialogue.gd since all it does is returning the Dictionary with the dialogue phrases, the actual building of the interface happens in DialogueBox. There, I fetch the textures by reaching to the database. 

Closes #92