MapActions created and grid removed from interactions
I used @nhydock's approach (sent via discord) to provide the correct nodes for the interactions aka using it's path. This implies on having to left click > edit children on every InteractivePawn after adding the desired actions. One alternative would be to provide the actions' scenes with the UI node, this way, each of them could display the UI when needed, without having to set node paths to them. Not sure which approach is better.
The actions are added to the InteractivePawn's `Interactions` node. I left one node in the map which starts an encounter after a dialogue, as mentioned by @NathanLovato
English mistakes fixed. 

Let me know if there are still problems with signals not being correctly connected.

Closes #50