Removed events from grid in local map (#104)
* Removed events from grid in local map

Closes #50

Also created boiler plates for pawns that have some type of interaction with the player. They work with both raycasts (in case we want something like pokemon) and area2ds for something that should only trigger on contact. 

* MapActions created and grid removed from interactions

I used @nhydock's approach (sent via discord) to provide the correct nodes for the interactions aka using it's path. This implies on having to left click > edit children on every InteractivePawn after adding the desired actions. One alternative would be to provide the actions' scenes with the UI node, this way, each of them could display the UI when needed, without having to set node paths to them. Not sure which approach is better.
The actions are added to the InteractivePawn's `Interactions` node. I left one node in the map which starts an encounter after a dialogue, as mentioned by @NathanLovato