Add an add-on to edit RectExtents interactively
Most of the work for #108

Currently only resizes uniformly around the rectangle's center! It needs
more code to add support for all 4 anchors. Or just fixing the issue
below, restoring the move tool's functionality and deselect, could do
the trick

RectExtents's internals changed a little, now it calculates its own
_rect property, a Rect2, based on its size and offset

Known issues: The input code blocks the move tool and selecting other
nodes at the moment

This is an exercise to get a sense for Godot's EditorPlugin API and
prepare some tutorial on the subject.

Big thanks to xphere, I used his godot-zone plugin as a reference to get
started :)