[WIP] Add new interface elements to represent the combat turn order...
This commit add two new scenes: TurnOrder and CombatPortrait.

The TurnOrder scene is designed to be a child of the CombatInterface node in
the CombatArena scene. It is initialized at the same time as the other
interface nodes, and dynamically display a CombatPortrait node for each
battlers, both PC and NPC.

For the moment, the CombatPortrait are just placeholders, and do not display
the real battlers portraits.