InteractivePawn triggers actions with player input
By getting near the pawn, a small balloon will show, allowing the player to press `ui_accept` to start the interaction. They can be started automatically by setting `auto_start_interaction` to `true`. Works with both Area2D triggers and the raycasts.

For the touch, it was already working as far as I could tell.

Closes #115