Fix #144: Player can move when interacting with NPCs
This relies on pausing the game: The NPCs still process during pause, as
e.g. if they were walking around you may not want to freeze the entire
game.

The PawnLeader stops as it inherits its pause mode from LocalMap and
GameBoard.

The StartCombatAction unpauses the game to play combat. I've changed the
interaction with map actions so they use the "finished" signal instead
of InteractivePawn waiting for the method to return.