Separate pawn animation from the PawnActor into a PawnAnim scene
Just like BattlerAnim before it, we have pawns that are reusable character
controllers, that apply to NPCs and to the playable character. And these pawns
could be any character in the game, so we need to be able to change the skin on
them.

Having separate scenes for animation is also essential for animators to be able
to do their job properly, while the devs can keep working on the Pawn scenes