Feat: allow to select cells when moving
Allow player to interrupt moving by selecting cells. However, the
player_cursor_marker will not appear until player_gamepieces move to a
cell. So there will be a noticeable delay of player_cursor_marker if
the speed of player_gamepieces is too slow or the size of tiles is too
big.